Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Argh, it looks like clang support of x87 in win64 may be intentional:
> https://forum.dlang.org/thread/qnyirdvfrzjnreediqjl@forum.dlang.org

Huh. Well that is unexpected. That link sort of suggests that x87 support is needed when targeting mingw (to support 80-bit long double) - I wonder what else it is used for, and if there is a switch we can throw to tell clang not to use the x87 FPU. Intel has been warning for a while that the x87 FPU is not recommended.

Patch-set: 4
CC: Gerrit User 1128439 <1128439@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>