vp9-svc: Fix issue with resettig lst_fb_idx under frame drops.
When encoding a given spatial layer, and the same spatial layer
on the previous superframe was dropped, we make sure the
lst_fb_idx for current layer corresponds to buffer slot
updated on last TL0 frame (same spatial layer).

This condition was to maintain proper temporal prediction pattern
under frame drops, and it should only apply to INTER frames.

But the condition was causing an assert to be triggered on spatial
layers whose base are key frames. Fix is to condition this reset of
lst_fb_idx on the is_key_frame flag.

Issue should only happen when superframe drop happens just before
a key frame.

Added datarate unittest the inserts key frame after superframe drop,
and verified that this fix is needed for test to pass.

Change-Id: I059d1135736aca17e1326b9b4a2b16371eb4634e