vp9-svc: Fix to pattern update when enabling layers on the fly.
When enabling spatial layer on the fly (say layer 2), that layer
may get a predictor to a buffer slot that is 2 layers below (layer 0).
This is checked and corrected for in vp9_svc_constrain_inter_layer_pred(),
but only when INTER_LAYER_PRED_ON_CONSTRAINED is set.

Fix is to always allow for this check (if inter_layer prediction is
used). A separate control should be used for setting
INTER_LAYER_PRED_ON_CONSTRAINED, as it conflicts with the
INTER_LAYER_PRED_ON/INTER_LAYER_PRED_OFF_NONKEY settings.

Change-Id: Ic6434000665f63aab27c509b5eb7b8fc965827bc