Update patch set 5
Patch Set 5:

It looks interesting and I like the idea. It would be nice to take it further and have a generic way of saying "run this N times". If it were me I'd probably try to do it with a macro but maybe there is a c++ class way that would work.

Patch-set: 5
Reviewer: Gerrit User 1117029 <1117029@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>