VP8: Fix use-after-free in postproc.
The pointer in vp8 postproc refers to show_frame_mi which is only
updated on show frame. However, when there is a no-show frame which also
changes the size (thus new frame buffers allocated), show_frame_mi is
not updated with new frame buffer memory.

Update show_frame_mi no matter the frame is no-show or not.

Bug: 842265
Change-Id: I33874f2112b39f74562cba528432b5f239e6a7bd