Update patch set 1
Patch Set 1:

Thanks for looking at this.

In terms of anything structural the other change in the patch is that the likelihood of a GF only group is higher though it could occur before in some situations.

Can we try reducing the max GF group size down even further to say 16 (which was a common default size limit for arf groups before) just to rule out any issue with group length.

If there is a problem with any encode that does not included arfs then this could impact low delay encodes as well.

Are the same clips failing as was the case before?

Patch-set: 1
Tag: mailMessageId=<CALRks=ROkJ2yjr2EJaLHj1qRzLTJ+UWV_k1L6gEgLY3F+=Co9Q@mail.gmail.com>