Force arf update for GF only groups.
In a sequence of GF only groups the ARF buffer pointer
continues to point at the previous KF or the last frame where
it was updated as part of an ARF group.

This patch forces the ARF buffer to be updated alongside the
GF buffer in GF only groups.

It is intended as a test case in regards to some HW decoders
that appear to have problems with very long GF groups on
static scenes but which also fail when the static section is
coded as multiple shorter GF only groups.

Change-Id: If46878003a6a3dc62b8e21a9e6e3fe6c4e9e7ad1