WIP: DO NOT SUBMIT
This cl should be broken up into multiple smaller cls.  The intent
is to hook some of these up to the sharpen parameter on command
line.  As is this is just to get comment...

1:  AQ MODE=1  Changes variance calculation to use avg of log of
4x4 variances. Instead of log of variance of the whole block. To
insure that variance values are more constant across different
partitioning sizes on different frames.

2: Greatly lowers rdmult used in trellis optimization

3: Shut off the end of block optimization that tries end of block
at every sub position if any of the coefficients are > 1.

4: Change the rounding factor to be exactly half.

5: Disable the skip block check that calculates RD using SSE from
predictor.

Change-Id: I4e1fe6024a370d0ae1d2fa6771b08f8578f28012