AQ MODE=1
Changes variance calculation to use avg of log of 4x4 variances.
Instead of log of variance of the whole block. To insure that
variance values are more constant across different
partitioning sizes on different frames.

Also changes so that segmentation is dependent on variance alone
not the difference between variance and avg frame variance.
Instead of changing segment id we change the effective rate ratio.

This should mean that the same block should maintain the same
segment id across frames even if lots of new stuff appears in
subsequent frames.

Change-Id: If275ceb7945b21968abdede170252a7443b7e9bc