Update patch set 1
Patch Set 1:

Johann: Thank you for the report. I forgot that libvpx supports pre-C99 compilers, and I am sorry for breaking the build.

I misunderstood your first comment -- I thought you reported that you already reverted this CL, which is why I did not revert it. I was planning to redo the fix in libaom first and then back-port the new fix to libvpx. It seems that Tom already wrote a new fix for libvpx. I will take a look and adopt it in libaom.

Patch-set: 1