Update patch set 9
Patch Set 9:

The implementation looks good. Regarding the performance test mentioned in the commit message, what is the number of tiles used in your test? What is the speed gain if number of threads is less than number of tiles(for example, 2 threads/4 tiles)? Is the synchronization overhead(checking if the row decoding is done) high in the multi-tile case?

Patch-set: 9
Reviewer: Gerrit User 1000884 <1000884@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>