Update patch set 24
Patch Set 24:

(1 comment)

I'm still struggling to see how can a buffer other than last, golden and/or alt be updated. Lets say, I want frame to use the last, golden and alt refs, which occupy buffers 0, 1 and 2 respectively, and I want this frame to be stored in buffer 3. Is that possible?

Patch-set: 24