vp9-svc: Fixes for cyclic refresh for SVC.
Add metrics that are being updated per-frame to
the layer struct, so each layer using the cyclic
refresh has the correct update.

And add flag to svc to indicate quality layers are
being used. Use it to allow for source_sad usage
for all layers when svc->quality_layers is on.

Change-Id: I0a9862fb6b6a79e894e2ff30c120dc4aa26fcda5