vp9: fix memory alloc for adaptive_rd_thresh_row_mt.
When the feature is enabled and the memory is not available, allocate
it. There was a case where speed feature changed in the middle of stream
but the number of tiles stayed the same, memory was not re-allocated.

Thus the memory for adaptive_rd_thresh_row_mt stayed NULL but the
feature was enabled.

Change-Id: I2f1f802ef98a554571b30094d3600b9439228457