Fix DoS in Loopfilter
While decoding an error clip with fewer threads than number of tiles,
one or more threads will decode more than one tile. It is possible in
this case that the first tile being decoded is erroneous. This will
cause the hook to exit without attempting to decode the second tile. In
this case, the vp9_set_row will not be called for the second tile.
Another thread that has started the loopfilter function would then wait
for the signal saying all the tiles have been decoded forever.

Change-Id: I7d61a35f518b67955cf9e2205d212c8fe834db82