Update patch set 2
Patch Set 2: Code-Review+2

It would makes sense I guess that that patch would cause the rate to be much lower in some cases and so I can see how it could be connected to the regression. There was a danger that the old code overshot its target by applying more bits to the ARF then also re-asserting the frame min rate a second time, but this could be addressed differently if need be, by not allowing the bits that a normal frame donates to the ARF to take it below the frame minimum.  This would limit the size of the ARF instead of allowing the rate used for normal frames (after donation to the ARF) to drop below the minimum.

For now I think we should apply this revert if it has been proven to be connected too the quality regression in some situations and then look to resolution of any impact on target vs output rate accuracy afterwards.

Patch-set: 2
Reviewer: Gerrit User 1000899 <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2