Update patch set 2
Patch Set 2:

Thinking on this overnight, one thing we do have to be aware of which was what triggered this change in the first place, is that while the revert wont do much harm in VBR cases where the min section % target is say <10%, where it is much higher this could cause substantial overshoot of what was expected. This happened and now would happen again for CVBR where initially I found that even for very easy content, if I had a min rate of 50% then the lowest I actually saw was close to 70 or 75%. Basically, we took away half the bits from the normal frames to add to the arf then added them back to push the frames rate back to its target minimum of 50%.

I am not sure what the min section % value is for the high value content for play movies but it is set to quite a high value for YT, though here it would currently be masked by the imposition of CQ. Hence this revert is certainly not a full answer and may as currently given cause the distributions in CVRB to be skewed away from the specified minimum. This could be countered by setting a lower minimum and by adjustment to the CVBR complexity number but is something we need to be aware of until we find a better way of handling this.

Simply not allowing as many bits to be given to the ARF may also be problematic in some cases as it may mean that those frames get starved too much.

Patch-set: 2