Update patch set 2
Patch Set 2:

Thanks. It should do because when using the 24 FPS source with the old code I see the problem in vpxenc as well. The reason it didn't happen with the converted 29.97 FPS version of the source is that the gap to that first key frame was long enough for their to be an arf group.

Patch-set: 2
Tag: mailMessageId=<CALRks=SdbKK7boxrR7LSa2+af+dR6XS790BAnR=Jw0QbyQ0qmA@mail.gmail.com>