Update patch set 1
Patch Set 1: Code-Review+1

I am OK with this, BUT ... another possibility is to extend / use what is already there.

The existing dual layer ARF experiment was enabled by setting --auto-alt-ref=2 (rather than 1 for normal 1 layer arf). Is there any reason this could not be used instead of adding a new control. In effect the current state is 0 (auto arf off) 1 (single layer), 2 (auto dual layer) and this could be extended to allow control not only of turning multi layer on (which I assume will replace completely the existing experiment as it is working better and is cleaner and more extensible.) but would also allow for passing in a limit on the maximum number of levels allowed.

Patch-set: 1
Reviewer: Gerrit User 1000899 <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1