Update patch set 1
Patch Set 1:

Note that increasing the factor here will tend to lengthen the group length even for very hard clips. This will hurt a lot for single layer but may help for multi-layer in some cases as it will allow multi-layer to be used where previously it might not be used because the initial single layer group length chosen was too short.  I think this is certainly a fertile avenue for investigation but as is this patch is problematic.

Patch-set: 1
Reviewer: Gerrit User 1000899 <1000899@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>