Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> What trouble are you having with that file? It might be best to open a bug for discussion:
> https://bugs.chromium.org/p/webm/issues/entry

The second frame header in the .ivf file indicates that the frame size is 2,684,590,080, but the allocated size for the buffer is 268,435,456. This causes the assert in ivf_video_source.h:81 to fail (Frame is too big for allocated code buffer). Even if I do increase the buffer size to what is indicated in the bitstream, there are only 113663 bytes left in the file, which causes the second assert in ivf_video_source.h:82 to fail (Failed to read complete frame).

Patch-set: 4
CC: Gerrit User 1289407 <1289407@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1289408 <1289408@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1289402 <1289402@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>