Update patch set 4
Patch Set 4:

Johann,
Sorry for the confusion here.
https://chromium-review.googlesource.com/c/webm/libvpx/+/1248221 added a file invalid-crbug-1562.ivf to invalid_file_test.cc.
This file is not yet uploaded to https://storage.googleapis.com/downloads.webmproject.org/test_data/libvpx/.
This resulted in Jenkins test failures on that commit.

In order to verify all Jenkins test pass, this commit was uploaded. We wanted to ensure all tests pass. This commit is not to be merged. 

Explanations given by Venkat are about why a different clip was used instead of the one used to report the issue.

Patch-set: 4