Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> (2 comments)

The latest patch set is just rebasing on HEAD.

I'm testing it on google3 autofuzz and it's hanging. There is still something wrong with handling of internal errors. 

I remembered the idea of this CL is to "copy" the behavior of VP9 (which has vpx_internal_error_info in MACROBLOCKD). So I added that. But an interesting behavior in VP9 is that the xd->error_info points to that in VP9_COMMON.

https://chromium.googlesource.com/webm/libvpx/+/master/vp9/common/vp9_onyxc_int.h#346

So they are sharing the vpx_internal_error_info. So what's the point of adding it to MACROBLOCKD if it's just pointing to that in VP9_COMMON (which is global across all threads)?

Patch-set: 7
Reviewer: Gerrit User 1171444 <1171444@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>