vp9: Reorganize the buffer level for cbr mode
Refactor the code with some changes.

Split update into two parts: move the fillup
(with per-frame-bandwidth) before the encoding, and
keep the leaking part (with encoded_frame_size) after
the encoding (postencde).

For SVC with ref_frame_config usage: use the timestamp
delta for the "framerate", instead of the framerate passed
in via the duration.

Adding the per-frame-bandwidth increase to the buffer in the
pre-encode causes some difference in performance
(since buffer level affects QP and frame-dropping), but change
is observed to be small.

Change-Id: I81a5562367034f318cffd451304bc4a34bf02a1d