Update patch set 1
Patch Set 1:

(2 comments)

This patch is correct but is it possible to switch it to size_t everywhere? Given that it refers to the size of the buffer that needs to be read in, that seems more correct to me.

Patch-set: 1
Reviewer: Gerrit User 1117029 <1117029@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>