subpixel_8t sse2: resolve missing declarations
vpx_asm_stubs.c only references these sse2 functions. Combine the files
similar to the way the ssse3/avx2 files are set up.

Mark the intrinsics as static because they are only used within the
macros here. It is unfortunate that the assembly functions can not be
marked static as well.

Change-Id: I342687a1046ae6ca46ae58644a7c170440de1dfb