Update patch set 3
Patch Set 3:

Yes, generally the RTCD system in libaom is the same as libvpx. In this particular instance the functions are subspecializations within a function. As they are never intended to be called directly, putting them in the RTCD does not really makes sense.

In this case, filter8_1dfunction is defined in a different header file so it still helps to keep the function prototypes in order.

Patch-set: 3
Reviewer: Gerrit User 1117029 <1117029@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>